fix type

+++ Libraries/Core/Models/Eft/Repair/TraderRepairActionDataRequest.cs
@@ -18,5 +18,5 @@ public record RepairItem
     public string? Id { get; set; }
 
     [JsonPropertyName("count")]
-    public int? Count { get; set; }
+    public double? Count { get; set; }
 }
